+++ lotus-commons/src/main/kotlin/util/IdGenerator.kt
@@ -2,7 +2,7 @@ package io.github.llh4github.lotus.commons.util
 
 import com.github.yitter.contract.IdGeneratorOptions
 import com.github.yitter.idgen.YitIdHelper
-import io.github.llh4github.lotus.commons.properties.IdGeneratorProperty
+import io.github.llh4github.lotus.commons.property.IdGeneratorProperty
 
 open class IdGenerator(property: IdGeneratorProperty) {
